+++ ruma-client-api/CHANGELOG.md
@@ -57,6 +57,7 @@ Breaking changes:
 * Move `r0::directory::{Filter, PublicRoomsChunk, RoomNetwork}` to
   the `ruma-common` crate
 * Replace `r0::room::create_room::InitialStateEvent` with `ruma_events::InitialStateEvent`
+* `error::ErrorKind` no longer implements `Copy`, `FromStr`
 
 Improvements: